// types/tarjetaConvenio.ts
export type TarjetaConvenio = {
    convenio_id: number; //Tabla Convenios
    empresa_id: number; //Tabla Empresas
    nombre_emp: string; //Tabla Empresas
    logo_url: string; //Tabla Empresas
    direccion: string; //Tabla Sucursales
    ciudad: string; //Tabla Sucursales
    categoria_id: number; //Tabla Categorias
    nombre_cat: string; //Tabla Categorias
    titulo: string; //Tabla Convenios
    condiciones: string; //Tabla Convenios
}